requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
httpx>=0.27.0
openai>=1.0.0
redis>=5.0.0
apscheduler>=3.10.0
tweepy>=4.14.0
kiwipiepy>=0.17.0
